V1.0.3 - Fixing emoji conditinal, .toLocaleString in the function and fixing one more return in the city's id
