+++ functions/src/controller/weather.controller.ts
@@ -1,9 +1,8 @@
 import { NextFunction, Request, Response } from "express";
-import { parse } from "fast-xml-parser";
 
-import { apiCptec } from "../config/apiCPTEC";
 import { City } from "../types/api.types";
 import { RequestDialog } from "../types/dialogflow.types";
+import { getCityId, getWeatherPrevision } from "../utils/api";
 
 export async function getWeather(
   req: Request,
@@ -12,35 +11,6 @@ export async function getWeather(
 ) {
   const body: RequestDialog = req.body;
   const city = body.queryResult.parameters.city;
-
-  async function getCityId(city: string) {
-    const data = apiCptec.get(`/listaCidades?city=${city}`).then((doc) => {
-      return doc;
-    });
-    const xml = (await (await data).data) as string;
-
-    const json = await parse(xml);
-
-    if (!json.cidades.cidade) {
-      return "false";
-    }
-
-    const id = json.cidades.cidade.id;
-    return id;
-  }
-
-  async function getWeatherPrevision(id: string) {
-    const data = apiCptec
-      .get(`/cidade/7dias/${id}/previsao.xml`)
-      .then((doc) => {
-        return doc;
-      });
-
-    const xml = (await (await data).data) as string;
-    const json = await parse(xml);
-
-    return json;
-  }
 
   const cityId = (await getCityId(city)) as string;
 
@@ -62,13 +32,13 @@ export async function getWeather(
 
   const { nome, atualizacao, previsao } = cidade;
 
-  const newDate = new Date(atualizacao).toLocaleDateString();
+  const newDate = new Date(atualizacao).toLocaleDateString("en-GB");
 
   const previsionList = previsao.map((doc) => {
     const newDate = new Date(doc.dia).toLocaleDateString();
     return `\r\nData: ${newDate} \r\nMaxima: ${doc.maxima}ºC ${
-      doc.maxima < "15" ? "🥶" : "🥵"
-    }  \r\nMinima: ${doc.minima}ºC ${doc.minima < "15" ? "🥶" : "🥵"} \r\n`;
+      doc.maxima <= "20" ? "🥶" : "🥵"
+    }  \r\nMinima: ${doc.minima}ºC ${doc.minima <= "20" ? "🥶" : "🥵"} \r\n`;
   });
 
   const str = await previsionList.toString().replace(/,/g, "");
